+++ ihc_tests/03_run/fup_001/a_execution_wrong_mixed_run.sh
@@ -1,0 +1,2 @@
+tko -m run cases.tio solver.py -u
+