+++ ihc_tests/03_run/fup_001/all_failures.sh
@@ -1,0 +1,1 @@
+tko -m run solver.py cases.tio -a