+++ ihc_tests/03_run/poo_002/side_run.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+tko -m run draft.ts cases.tio -s