+++ ihc_tests/03_run/poo_002/simple_run.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-tko -m run draft.ts cases.tio