+++ ihc_tests/05_loading_files/missing_target.sh
@@ -1,0 +1,14 @@
+#!/bin/bash
+
+tko -m run
+echo ""
+
+tko -m run cases.t
+echo ""
+
+tko -m run solver.py cases.t
+echo ""
+
+tko -m run solver.p cases.tio
+echo ""
+