+++ ihc_tests/05_loading_files/running_modes.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+tko -m run cases.tio