+++ ihc_tests/05_loading_files/running_modes2.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+
+tko -m run solver.py