+++ ihc_tests/05_loading_files/running_modes4.sh
@@ -1,0 +1,2 @@
+
+tko -m run solver.py cases.tio