+++ src/tko/__main__.py
@@ -17,34 +17,30 @@ from .__init__ import __version__
 
 
 class Main:
-    @staticmethod
-    def execute(args):
-        Actions.exec(args.target_list)
 
     @staticmethod
     def run(args):
-        if args.width is not None:
-            Report.set_terminal_size(args.width)
         PatternLoader.pattern = args.pattern
         param = Param.Basic().set_index(args.index)
         if args.quiet:
             param.set_diff_mode(DiffMode.QUIET)
-        if args.vertical:
+        elif args.first:
+            param.set_diff_mode(DiffMode.FIRST)
+        elif args.all:
+            param.set_diff_mode(DiffMode.ALL)
+
+
+        # load default diff from settings if not specified
+        if not args.sideby and not args.updown:
+            param.set_up_down(not SettingsParser().get_hdiff())
+        elif args.sideby:
+            param.set_up_down(False)
+        elif args.updown:
             param.set_up_down(True)
         Actions.run(args.target_list, param)
 
     @staticmethod
-    def list(args):
-        if args.width is not None:
-            Report.set_terminal_size(args.width)
-        PatternLoader.pattern = args.pattern
-        param = Param.Basic().set_index(args.index)
-        Actions.list(args.target_list, param)
-
-    @staticmethod
     def build(args):
-        if args.width is not None:
-            Report.set_terminal_size(args.width)
         PatternLoader.pattern = args.pattern
         manip = Param.Manip().set_unlabel(args.unlabel).set_to_sort(args.sort).set_to_number(args.number)
         Actions.build(args.target, args.target_list, manip, args.force)
@@ -78,9 +74,9 @@ class Main:
     #     Actions.update(args.target_list, manip, args.cmd)
     #     return 0
 
-    @staticmethod
-    def update(_args):
-        Down.update()
+    # @staticmethod
+    # def update(_args):
+    #     Down.update()
 
     @staticmethod
     def down(args):
@@ -90,7 +86,6 @@ class Main:
     @staticmethod
     def main():
         parent_basic = argparse.ArgumentParser(add_help=False)
-        parent_basic.add_argument('--width', '-w', type=int, help="term width")
         parent_basic.add_argument('--index', '-i', metavar="I", type=int, help='run a specific index.')
         parent_basic.add_argument('--pattern', '-p', metavar="P", type=str, default='@.in @.sol',
                                   help='pattern load/save a folder, default: "@.in @.sol"')
@@ -104,29 +99,28 @@ class Main:
                                   help='pattern load/save a folder, default: "@.in @.sol"')
 
         parser = argparse.ArgumentParser(prog='tko', description='A tool for competitive programming.')
-        parser.add_argument('-v', '--version', action='store_true', help='show version.')
-        parser.add_argument('-g', '--guide', action='store_true', help='show simple tko guide.')
-        parser.add_argument('-b', '--bash', action='store_true', help='show simple bash guide.')
-        parser.add_argument('-m', '--mono', action='store_true', help='monochromatic.')
-        parser.add_argument('-c', '--config', type=str, help='config file.')
+        parser.add_argument('-c', metavar='CONFIG_FILE', type=str, help='config file.')
+        parser.add_argument('-w', metavar='WIDTH', type=int, help="terminal width.")
+        parser.add_argument('-v', action='store_true', help='show version.')
+        parser.add_argument('-g', action='store_true', help='show tko simple guide.')
+        parser.add_argument('-b', action='store_true', help='show bash simple guide.')
+        parser.add_argument('-m', action='store_true', help='monochromatic.')
         
         subparsers = parser.add_subparsers(title='subcommands', help='help for subcommand.')
-
-        # list
-        parser_l = subparsers.add_parser('list', parents=[parent_basic], help='show test entries.')
-        parser_l.add_argument('target_list', metavar='T', type=str, nargs='*', help='targets.')
-        parser_l.set_defaults(func=Main.list)
-
-        # exec
-        parser_e = subparsers.add_parser('exec', parents=[parent_basic], help='run without test.')
-        parser_e.add_argument('target_list', metavar='T', type=str, nargs='*', help='target.')
-        parser_e.set_defaults(func=Main.execute)
 
         # run
         parser_r = subparsers.add_parser('run', parents=[parent_basic], help='run with test cases.')
         parser_r.add_argument('target_list', metavar='T', type=str, nargs='*', help='solvers, test cases or folders.')
-        parser_r.add_argument('--vertical', '-v', action='store_true', help="use vertical mode.")
-        parser_r.add_argument('--quiet', '-q', action='store_true', help='quiet mode, do not show diffs')
+        
+        group_n = parser_r.add_mutually_exclusive_group()
+        group_n.add_argument('--first', '-f', action='store_true', help='(default) show only first failure.')
+        group_n.add_argument('--quiet', '-q', action='store_true', help='quiet mode, do not show any failure.')
+        group_n.add_argument('--all', '-a', action='store_true', help='show all failures.')
+
+        # add a exclusive group for diff mode
+        group = parser_r.add_mutually_exclusive_group()
+        group.add_argument('--updown', '-u', action='store_true', help="diff mode up-to-down.")
+        group.add_argument('--sideby', '-s', action='store_true', help="diff mode sidebyside.")
         parser_r.set_defaults(func=Main.run)
 
         # build
@@ -138,14 +132,10 @@ class Main:
 
         # down
         parser_d = subparsers.add_parser('down', help='download problem from repository.')
-        parser_d.add_argument('disc', type=str, help=" [ fup | ed | poo ]")
-        parser_d.add_argument('index', type=str, help="3 digits label like 021")
+        parser_d.add_argument('disc', type=str, nargs='?', help=" [ fup | ed | poo ].")
+        parser_d.add_argument('index', type=str, nargs='?', help="3 digits label like 021.")
         parser_d.add_argument('extension', type=str, nargs='?', default="-", help="[ c | cpp | js | ts | py | java ]")
         parser_d.set_defaults(func=Main.down)
-
-        # update
-        parser_u = subparsers.add_parser('update', help='update problem from repository.')
-        parser_u.set_defaults(func=Main.update)
 
         # settings
         parser_s = subparsers.add_parser('config', help='settings tool.')
@@ -156,29 +146,33 @@ class Main:
 
         args = parser.parse_args()
 
-        if args.config:
-            SettingsParser().set_settings_file(args.config)
-
         if len(sys.argv) == 1:
             parser.print_help()
             return
-        
+
+
+        # setting general settings options
+        if args.w is not None:
+            Report.set_terminal_size(args.width)
+
+        if args.c:
+            SettingsParser().set_settings_file(args.config)
         
         if SettingsParser().get_ascii():
             symbols.set_ascii()
         else:
             symbols.set_unicode()
 
-        if not args.mono and SettingsParser().get_color():
+        if not args.m and SettingsParser().get_color():
             Colored.enabled = True
             symbols.set_colors()
 
-        if args.version or args.guide or args.bash:
-            if args.version:
+        if args.v or args.g or args.b:
+            if args.v:
                 print("tko version " + __version__)
-            if args.bash:
+            if args.b:
                 print(bash_guide[1:], end="")
-            if args.guide:
+            if args.g:
                 print(tko_guide[1:], end="")
         else:
             try: